refactor: create a dedicated package for the YAML syntax descriptors
